+++ cmd/dbaas/dbaas_user_list.go
@@ -67,6 +67,8 @@ func (c *dbaasUserListCmd) CmdRun(cmd *cobra.Command, args []string) error {
 		return c.listValkey(cmd, args)
 	case "thanos":
 		return c.listThanos(cmd, args)
+	case "clickhouse":
+		return c.listClickhouse(cmd, args)
 	default:
 		return fmt.Errorf("listing users unsupported for service of type %q", db.Type)